Refactor estimateTotalGasRequiredByUnsignedTransaction into QueryBase
